move memStore
